Move ApiHelpers.test.jsx
Move NetworkGraph.test.jsx
