+++ web/app/js/components/util/ApiHelpers.test.jsx
@@ -2,11 +2,11 @@
 import _ from 'lodash';
 import 'raf/polyfill'; // the polyfill import must be first
 import Adapter from 'enzyme-adapter-react-16';
-import ApiHelpers from '../js/components/util/ApiHelpers.jsx';
+import ApiHelpers from './ApiHelpers.jsx';
 import Enzyme from 'enzyme';
 import { expect } from 'chai';
 import { mount } from 'enzyme';
-import { routerWrap } from './testHelpers.jsx';
+import { routerWrap } from '../../../test/testHelpers.jsx';
 import sinon from 'sinon';
 import sinonStubPromise from 'sinon-stub-promise';
 /* eslint-enable */
